move `make_mol` over
