+++ chemprop/v2/utils/__init__.py
@@ -1,2 +1,2 @@
 from .mixins import RegistryMixin
-from .utils import AutoName
+from .utils import AutoName, make_mol
